+++ duckscript_sdk/src/sdk/std/env/set_env/mod.rs
@@ -1,5 +1,9 @@
 use crate::utils::pckg;
-use duckscript::types::command::{Command, CommandResult};
+use crate::utils::state::{get_as_string, get_handles_sub_state};
+use duckscript::types::command::{Command, CommandResult, Commands};
+use duckscript::types::instruction::Instruction;
+use duckscript::types::runtime::StateValue;
+use std::collections::HashMap;
 use std::env;
 
 #[cfg(test)]
@@ -28,15 +32,52 @@ impl Command for CommandImpl {
         Box::new((*self).clone())
     }
 
-    fn run(&self, arguments: Vec<String>) -> CommandResult {
+    fn requires_context(&self) -> bool {
+        true
+    }
+
+    fn run_with_context(
+        &self,
+        arguments: Vec<String>,
+        state: &mut HashMap<String, StateValue>,
+        _variables: &mut HashMap<String, String>,
+        _output_variable: Option<String>,
+        _instructions: &Vec<Instruction>,
+        _commands: &mut Commands,
+        _line: usize,
+    ) -> CommandResult {
         if arguments.is_empty() {
             CommandResult::Error("Missing environment variable name and value.".to_string())
         } else if arguments.len() == 1 {
             CommandResult::Error("Missing environment variable value.".to_string())
         } else {
-            env::set_var(&arguments[0], &arguments[1]);
+            if arguments[0] == "--handle" {
+                let state = get_handles_sub_state(state);
 
-            CommandResult::Continue(Some("true".to_string()))
+                let key = &arguments[1];
+
+                match state.get(key) {
+                    Some(state_value) => match state_value {
+                        StateValue::SubState(map) => {
+                            for (env_key, env_value) in map {
+                                if let Ok(env_value_string) = get_as_string(env_value) {
+                                    env::set_var(&env_key, &env_value_string);
+                                }
+                            }
+
+                            CommandResult::Continue(Some("true".to_string()))
+                        }
+                        _ => CommandResult::Error("Invalid handle provided.".to_string()),
+                    },
+                    None => CommandResult::Error(
+                        format!("Map for handle: {} not found.", key).to_string(),
+                    ),
+                }
+            } else {
+                env::set_var(&arguments[0], &arguments[1]);
+
+                CommandResult::Continue(Some("true".to_string()))
+            }
         }
     }
 }