+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * New join_path command.
 * New glob_cp command #192
+* set_env command now accepts a map handle and sets all env vars from the map
 
 ### v0.8.4 (2021-21-07)